Optionally locate Druid tasks through "runningTasks" endpoint on the overlord.
This is enabled by setting druidBeam.taskLocator=overlord, and works by polling the
overlord in a background thread. Also added a useless setting for
druidBeam.overlordLocator, which might not be useless in the future if we ever have
a way to do that without Curator discovery (perhaps a hard-coded list, or a DNS name).

No integration tests yet, since the Druid code that enables this is not yet in a
Maven release so we cannot pull it in.